+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `create-element-extension`: Warns when using the second parameter of `createElement` for element extension using the `is` attribute.
 - `validate-element-name`: Warns when using an invalid element name or when there are potential issues or conflicts with the element name.
 - `root-selector-to-html`: Warns when using the `:root` selector in `custom-style` or element styles.
+- `custom-style-extension`: Warns when `<style>` is extended by `is="custom-style"` but not wrapped by `<custom-style>`.
 
 ### Fixes
 - `content-selector-to-slotted`: Don't attempt to fix usages where a selector would come after `::slotted`